+++ pmd-core/src/main/java/net/sourceforge/pmd/RuleSetLoader.java
@@ -134,14 +134,10 @@ public final class RuleSetLoader {
     }
 
     /**
-     * Create a new rule set factory, if you have to (that class is deprecated).
+     * Create a new rule set factory, if you have to (that class is internal).
      * That factory will use the configuration that was set using the setters of this.
-     *
-     * @deprecated {@link RuleSetFactory} is deprecated, replace its usages
-     *     with usages of this class, or of static factory methods of {@link RuleSet}
      */
-    @Deprecated
-    public RuleSetFactory toFactory() {
+    RuleSetFactory toFactory() {
         return new RuleSetFactory(
             this.resourceLoader,
             this.languageRegistry,